x64 on windows compilable (but not linkable yet), some warnings fixed

+++ src/platform-win/fontgen.win.cpp
@@ -95,7 +95,7 @@ FontGeneratorResult FontGenerator::generate(ImageFormat fmt)
 		float lineHeight = 0.0f;
 		float y_pos = characterOffset;
 		x_pos = characterOffset;
-		textureSize.x = roundToHighestPowerOfTwo(static_cast<size_t>(textureWidth));
+		textureSize.x = static_cast<int>(roundToHighestPowerOfTwo(static_cast<size_t>(textureWidth)));
 		textureWidth = static_cast<float>(textureSize.x);
 		for (CharDescriptorList::iterator i = chars.begin(), e = chars.end(); i != e; ++i)
 		{
@@ -110,7 +110,7 @@ FontGeneratorResult FontGenerator::generate(ImageFormat fmt)
 			x_pos += i->size.x + characterOffset;
 			lineHeight = etMax(i->size.y, lineHeight);
 		}
-		textureSize.y = roundToHighestPowerOfTwo( static_cast<size_t>(y_pos + lineHeight + characterOffset) );
+		textureSize.y = static_cast<int>(roundToHighestPowerOfTwo(static_cast<size_t>(y_pos + lineHeight + characterOffset)));
 
 		foundOptimal = textureSize.y / textureSize.x <= 1;
